+++ src/js/module/dashboard/ListIssuesEngineering.js
@@ -67,6 +67,7 @@ class ListIssuesEngineering extends React.Component {
                 extraClass="none"
                 item="issue"
                 data={this.props.issues}
+                applyFilters
             />
         );
     }