+++ src/js/module/K2picker/K2picker.js
@@ -1,8 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import ReactNativeOnyx from 'react-native-onyx';
 import K2PickerPicker from './K2PickerPicker';
-import ONYXKEYS from '../../ONYXKEYS';
 
 /**
  * Displays the picker for the KSV2 Labels
@@ -12,9 +10,6 @@ import ONYXKEYS from '../../ONYXKEYS';
 export default function () {
     return {
         draw() {
-            ReactNativeOnyx.init({
-                keys: ONYXKEYS,
-            });
             ReactDOM.render(
                 <K2PickerPicker />,
                 document.getElementsByClassName('k2picker-wrapper')[0],