+++ src/js/module/ReviewedDocComment/ReviewedDocComment.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import ReviewedDocCommentButton from './ReviewedDocCommentButton';
+
+export default function () {
+    return {
+        draw() {
+            ReactDOM.render(
+                <ReviewedDocCommentButton />,
+                document.getElementsByClassName('k2revieweddoccomment-wrapper')[0],
+            );
+        },
+    };
+}